+++ app/src/main/java/com/group147/appartmentblog/repositories/PostRepository.kt
@@ -1,7 +1,7 @@
 package com.group147.appartmentblog.repositories
 
 import androidx.lifecycle.LiveData
-import com.group147.appartmentblog.dao.PostDao
+import com.group147.appartmentblog.database.post.PostDao
 import com.group147.appartmentblog.model.Post
 
 class PostRepository(private val postDao: PostDao) {
